fix: redundant cast

+++ src/creatures/players/player.cpp
@@ -6785,7 +6785,7 @@ void Player::checkSkullTicks(int64_t ticks) {
 void Player::updateBaseSpeed() {
 	const uint16_t maxSpeed = hasFlag(PlayerFlags_t::SetMaxSpeed) ? PLAYER_MAX_STAFF_SPEED : PLAYER_MAX_SPEED;
 	if (!hasFlag(PlayerFlags_t::SetMaxSpeed)) {
-		const uint32_t computedSpeed = static_cast<uint32_t>(vocation->getBaseSpeed()) + (level - 1);
+		const uint32_t computedSpeed = vocation->getBaseSpeed() + (level - 1);
 		baseSpeed = static_cast<uint16_t>(std::min<uint32_t>(computedSpeed, maxSpeed));
 	} else {
 		baseSpeed = maxSpeed;
